requirements.txt
# Core Framework
chainlit>=2.5.0

# LangChain Ecosystem
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-chroma>=0.2.0
langchain-community>=0.3.0
langchain-huggingface>=0.3.0

# Vector Database
chromadb>=1.0.0
lark>=1.0.0

# AI/ML APIs
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0


# Evaluation
ragas>=0.1.0
datasets>=2.0.0
